list ids, more tests, improved sql
Added test fixtures for connections to local databases that are
either completely empty or have a small number of dogs, to prevent
tests from assuming that there are dogs in the database (or at least
to make explicit which do). Also, added sql teardown script to reset
the database to empty (if you run teardown and then setup). Updated
cloudfunction for legibility and to mandate schemas for any input/
output.

Authored-by: Julian Zucker <julian.zucker@gmail.com>